cleanup: fix warnings

+++ src/graphics/fonts/psf.c
@@ -6,16 +6,16 @@
 
 void psf2_draw_char(psf2_header_t *font, struct limine_framebuffer *framebuffer, char c, int x, int y, uint32_t fg, uint32_t bg) {
         unsigned char *glyph = (unsigned char *)font + font->header_size +
-                (c > 0 && c < font->glyph_count ? c : 0) * font->glyph_size;
+                (c > 0 && (uint32_t)c < font->glyph_count ? c : 0) * font->glyph_size;
 
         int bytes_per_glyph_line = (font->width + 7) / 8;
         int row_offset = y * framebuffer->pitch;
 
-        for (int ty = 0; ty < font->height; ty++) {
+        for (uint32_t ty = 0; ty < font->height; ty++) {
                 unsigned char *current_byte = glyph + ty * bytes_per_glyph_line;
                 uint8_t mask = 1 << 7;
 
-                for (int tx = 0; tx < font->width; tx++) {
+                for (uint32_t tx = 0; tx < font->width; tx++) {
                         uint32_t *pixel = (uint32_t *)((uint8_t *)framebuffer->address + row_offset + (x + tx) * 4);
                         *pixel = (*current_byte & mask) ? fg : bg;
